requirements.txt
djangorestframework-simplejwt
mysqlclient
django-environ
requests